requirements.txt
Flask==1.1.4
Werkzeug==1.0.1
mysql-connector-python
flask-restful
psycopg2-binary
passlib
email-validator
flask-jwt-extended
